// src/resource/enum/types.ts
export enum Type {
  Int = "int",
  String = "string",
  Boolean = "boolean",
  Seq = "seq",
  Locshape = "locshape",
  Component = "component",
  Idk = "idk",
  Midi = "midi",
  NpcMode = "npc_mode",
  Namedobj = "namedobj",
  Synth = "synth",
  Stat = "stat",
  NpcStat = "npc_stat",
  Fontmetrics = "fontmetrics",
  Enum = "enum",
  Loc = "loc",
  Model = "model",
  Npc = "npc",
  Obj = "obj",
  PlayerUid = "player_uid",
  Spotanim = "spotanim",
  NpcUid = "npc_uid",
  Inv = "inv",
  Category = "category",
  Struct = "struct",
  Dbrow = "dbrow",
  Interface = "interface",
  Dbtable = "dbtable",
  Coord = "coord",
  Mesanim = "mesanim",
  Param = "param",
  Queue = "queue",
  Weakqueue = "weakqueue",
  Timer = "timer",
  Softtimer = "softtimer",
  Char = "char",
  Dbcolumn = "dbcolumn",
  Proc = "proc",
  Label = "label",
  Var = "var",
  Hunt = "hunt",
  Walktrigger = "walktrigger",
  Ob2 = "ob2",
  Idkit = "idkit",
  Frame = "frame"
}